requirements.txt
Flask==3.0.0
opencv-python-headless==4.8.1.78
pytesseract==0.3.10
Pillow==10.1.0
pandas==2.1.4
numpy==1.25.2
gunicorn==21.2.0
python-multipart==0.0.6